ETL/1_carga_datos_crudos/requirements_api.txt
functions-framework==3.*
google-cloud-storage
googlemaps
pandas
pyarrow
fastparquet